moved ddtrace.ext.priority constants
