+++ ddtrace/sampler.py
@@ -11,12 +11,12 @@ from typing import TYPE_CHECKING
 
 import six
 
+from .constants import AUTO_KEEP
+from .constants import AUTO_REJECT
 from .constants import ENV_KEY
 from .constants import SAMPLING_AGENT_DECISION
 from .constants import SAMPLING_LIMIT_DECISION
 from .constants import SAMPLING_RULE_DECISION
-from .ext.priority import AUTO_KEEP
-from .ext.priority import AUTO_REJECT
 from .internal.compat import iteritems
 from .internal.compat import pattern_type
 from .internal.logger import get_logger
