+++ releasenotes/notes/moved-priority-constants-316d6b3b791d6b3e.yaml
@@ -1,0 +1,4 @@
+---
+deprecations:
+  - |
+    Moved `ddtrace.ext.priority` constants into `ddtrace.constants` module.
